Re-hash legacy plain-text passwords and stop attempt ids from wrapping

+++ api/src/main/java/ua/nure/latysh/quizzes/api/attempt/AttemptCompletionResponse.java
@@ -2,6 +2,6 @@ package ua.nure.latysh.quizzes.api.attempt;
 
 import java.time.Instant;
 
-public record AttemptCompletionResponse(long attemptId, int quizId, int score, Instant completedAt) {
+public record AttemptCompletionResponse(int attemptId, int quizId, int score, Instant completedAt) {
 }
